added javadocs and refactored some files
