+++ src/app/AnalyzePlaylistUseCaseFactory.java
@@ -6,8 +6,8 @@ import interface_adapter.analyze_playlist.AnalyzePlaylistPresenter;
 import interface_adapter.analyze_playlist.AnalyzePlaylistViewModel;
 import use_case.analyze_playlist.*;
 import use_case.login.LoginDataAccessInterface;
-import use_case.login.services.TracksAudioFeaturesAPIHandler;
-import use_case.login.services.UserPlaylistItemsAPIHandler;
+import use_case.services.TracksAudioFeaturesAPIHandler;
+import use_case.services.UserPlaylistItemsAPIHandler;
 import view.AnalyzePlaylistView;
 
 /**
