+++ src/app/CreatePlaylistUseCaseFactory.java
@@ -5,7 +5,6 @@ import interface_adapter.create_playlist.CreatePlaylistController;
 import interface_adapter.create_playlist.CreatePlaylistPresenter;
 import interface_adapter.create_playlist.CreatePlaylistViewModel;
 import interface_adapter.playlist_created.PlaylistCreatedViewModel;
-import interface_adapter.view_moods.ViewMoodsViewModel;
 import use_case.create_mood.MoodDataAccessInterface;
 import use_case.create_playlist.CreatePlaylistInputBoundary;
 import use_case.create_playlist.CreatePlaylistInteractor;
@@ -14,11 +13,21 @@ import use_case.services.CreatePlaylistAPIHandler;
 import use_case.services.GetRecommendationAPIHandler;
 import use_case.login.LoginDataAccessInterface;
 import use_case.services.UserTopTracksAPIHandler;
-import use_case.services.UserTopTracksAPIHandlerInterface;
 import view.CreatePlaylistView;
 
+/**
+ * This class is a Factory for functions that create aspects of the CreatePlaylist use case
+ */
 public class CreatePlaylistUseCaseFactory {
-    private CreatePlaylistUseCaseFactory() {}
+
+    /**
+     * Create the View for the CreateMood Screen
+     * @param viewManagerModel The view manager model that handles the views
+     * @param createPlaylistViewModel The view model for the CreateMood View
+     * @param playlistCreatedViewModel The view model for the ViewMoods view
+     * @param moodDataAccessObject The data access object that deals with the Moods stored
+     * @return
+     */
 
     public static CreatePlaylistView create(ViewManagerModel viewManagerModel,
                                             CreatePlaylistViewModel createPlaylistViewModel,
