+++ src/interface_adapter/create_playlist/CreatePlaylistPresenter.java
@@ -5,6 +5,9 @@ import interface_adapter.playlist_created.PlaylistCreatedState;
 import interface_adapter.playlist_created.PlaylistCreatedViewModel;
 import use_case.create_playlist.CreatePlaylistOutputBoundary;
 
+/**
+ * The Presenter for the Create Playlist use case
+ */
 public class CreatePlaylistPresenter implements CreatePlaylistOutputBoundary{
 
     private final CreatePlaylistViewModel createPlaylistViewModel;
@@ -13,6 +16,12 @@ public class CreatePlaylistPresenter implements CreatePlaylistOutputBoundary{
 
     private ViewManagerModel viewManagerModel;
 
+    /**
+     * The constructor for the Create Playlist Presenter object
+     * @param createPlaylistViewModel The view model for the Create Playlist View
+     * @param playlistCreatedViewModel The view model for the Playlist Created View
+     * @param viewManagerModel The view manager to handle the views
+     */
     public CreatePlaylistPresenter(CreatePlaylistViewModel createPlaylistViewModel, PlaylistCreatedViewModel playlistCreatedViewModel,
                                    ViewManagerModel viewManagerModel) {
         this.createPlaylistViewModel = createPlaylistViewModel;
@@ -20,6 +29,9 @@ public class CreatePlaylistPresenter implements CreatePlaylistOutputBoundary{
         this.viewManagerModel = viewManagerModel;
     }
 
+    /**
+     * Prepares a success screen
+     */
     @Override
     public void prepareSuccessView() {
         System.out.println("Playlist successfully created");
@@ -30,6 +42,10 @@ public class CreatePlaylistPresenter implements CreatePlaylistOutputBoundary{
         this.viewManagerModel.firePropertyChanged();
     }
 
+    /**
+     * Triggers the pop up window on failure of the use case
+     * @param error The error that caused the failure
+     */
     @Override
     public void prepareFailView(String error) {
         CreatePlaylistState createPlaylistState = createPlaylistViewModel.getState();
