+++ src/interface_adapter/logged_in/LoggedInViewModel.java
@@ -2,8 +2,6 @@ package interface_adapter.logged_in;
 
 import interface_adapter.ViewModel;
 
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
 
@@ -15,7 +13,7 @@ public class LoggedInViewModel extends ViewModel {
 
     public final String LOGOUT_BUTTON_LABEL = "Log out";
     public final String CREATE_MOOD_BUTTON_LABEL = "Create Custom Mood";
-    public final String GET_PLAYLIST_BUTTON_LABEL = "Get Custom Playlist";
+    public final String CREATE_PLAYLIST_BUTTON_LABEL = "Create Custom Playlist";
     public final String ANALYZE_PLAYLIST_BUTTON_LABEL = "Analyze Playlist";
     public final String GROUP_PLAYLIST_BUTTON_LABEL = "Generate Group Playlist";
