+++ src/use_case/services/UserPlaylistItemsAPIHandlerInterface.java
@@ -1,4 +1,4 @@
-package use_case.login.services;
+package use_case.services;
 
 import java.io.IOException;
 import java.util.List;
